requirements.txt
# Airflow providers
apache-airflow-providers-docker==4.4.3

# Data
pandas==2.2.3
pyarrow==17.0.0

# Data Quality
great-expectations==1.7.1

# ML
mlflow==2.16.0
scikit-learn==1.5.2
numpy==2.1.3
matplotlib==3.9.2

# S3/MinIO (optional for later)
boto3==1.35.54
s3fs==2024.10.0
python-dateutil==2.9.0.post0
